fix(gig): rename skill to workerSkill
