+++ libs/product-domain/backend/src/gig/infrastructure/drizzle/repositories/index.ts
@@ -3,7 +3,7 @@ export * from './gig-work';
 export * from './gig-work-team';
 export * from './recommendation';
 export * from './review';
-export * from './skill';
+export * from './worker-skill';
 export * from './skill-hire';
 export * from './slot';
 export * from './worker';
