+++ libs/product-domain/backend/src/gig/infrastructure/drizzle/repositories/worker-skill.ts
