+++ libs/product-domain/backend/src/gig/infrastructure/drizzle/schemas/gig-db/index.ts
@@ -4,6 +4,6 @@ export * from './gig-work-teams';
 export * from './recommendations';
 export * from './review';
 export * from './skill-hires';
-export * from './skills';
+export * from './worker-skills';
 export * from './slots';
 export * from './workers';
