+++ libs/product-domain/backend/src/gig/infrastructure/drizzle/schemas/gig-db/worker-skills.ts
